Rename frame source files to match Layer["type"], extract image.
