+++ src/util.ts
@@ -5,6 +5,8 @@ import { pathExists } from 'fs-extra';
 import type { Keyframe } from './types.js';
 import type { Position, PositionObject, Transition } from './types.js';
 import type { TOriginX, TOriginY } from 'fabric';
+import * as fabric from 'fabric/node';
+import fileUrl from 'file-url';
 
 export function toArrayInteger(buffer: Buffer) {
   if (buffer.length > 0) {
@@ -122,3 +124,5 @@ export const assertFileValid = async (path: string, allowRemoteRequests?: boolea
 export function checkTransition(transition?: Transition | null) {
   assert(transition == null || typeof transition === 'object', 'Transition must be an object');
 }
+
+export const loadImage = (pathOrUrl: string) => fabric.util.loadImage(isUrl(pathOrUrl) ? pathOrUrl : fileUrl(pathOrUrl));
